Switch config file format from INI to PHP.
Since the properties file has been deprecated as a way to configure
log4php, I switched the config file used here to the native php config
format.  I figure this is preferable for two reasons over the XML style,
one is that CI uses php for all it's other configs, and two is that the
XML configs require an extra step internally to be converted to the
native PHP array format.

Doing this uncovered a small bug in the Logger.php class as well where
the '.php' extension was not supported by the method that loads the
configurator.  I patched this also.

http://logging.apache.org/log4php/docs/configuration.html